#32374954: Display messages
A state object has been added to the server which will keep the messages stored in memory. Once a client has connected to the server the entire state will be sent to the client which can then display the messages.

Also added prop-types to increase readability and to enable clear error messages if a prop has for example the wrong type or is missing.

Story: https://github.com/alexdarland/chat-app/projects/1#card-32374954