#32374950: Post message
Client is now able to post messages to the server which is stored in-memory. An event is also fired to all connected clients and the new message will be appended to the list of messages on screen.

Also included minor and relevant refactoring.

Story: https://github.com/alexdarland/chat-app/projects/1#card-32374950